+++ homeworks/fileshare/src/main/java/com/example/fileshare/service/FileServiceImpl.java
@@ -6,6 +6,7 @@ import com.example.fileshare.repository.FileRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -26,34 +27,47 @@ public class FileServiceImpl implements FileService {
 
     @Override
     public void delete(User user, int id) {
-        File file = fileRepository.findById(id);
-        if(file.getAuthor().equals(user)){
-            fileRepository.delete(file);
+        Optional<File> optionalFile = fileRepository.findById(id);
+        if(optionalFile.isPresent()){
+            File file = optionalFile.get();
+
+            if(file.getAuthor().equals(user)){
+                fileRepository.delete(file);
+            }
         }
     }
 
     @Override
     public void updateName(User user, int id, String newName) {
-        File file = fileRepository.findById(id);
-        if(file.getAuthor().equals(user)){
-            file.setName(newName);
-            fileRepository.save(file);
+        Optional<File> optionalFile = fileRepository.findById(id);
+        if(optionalFile.isPresent()){
+            File file = optionalFile.get();
+
+            if(file.getAuthor().equals(user)){
+                file.setName(newName);
+                fileRepository.save(file);
+            }
         }
     }
 
     @Override
     public void updateParent(User user, int id, File parent) {
-        File file = fileRepository.findById(id);
-        if(file.getAuthor().equals(user)){
-            file.setParent(parent);
-            fileRepository.save(file);
+        Optional<File> optionalFile = fileRepository.findById(id);
+        if(optionalFile.isPresent()){
+            File file = optionalFile.get();
+
+            if(file.getAuthor().equals(user)){
+                file.setParent(parent);
+                fileRepository.save(file);
+            }
         }
+
     }
 
     @Override
     public List<File> findRootFiles(User user) {
         return fileRepository
-                .findAllByAuthor(user)
+                .findAllByAuthor(user.getId())
                 .stream()
                 .filter(file -> file.parent == null)
                 .collect(Collectors.toList());
@@ -61,12 +75,13 @@ public class FileServiceImpl implements FileService {
 
     @Override
     public List<File> findFiles(User user, File parent) {
-        return fileRepository.findAllByAuthorAndParent(user, parent);
+        return fileRepository.findAllByAuthorAndParent(user.getId(), parent.getId());
     }
 
 
     @Override
     public File findFileById(int id) {
-        return fileRepository.findById(id);
+        Optional<File> optionalFile = fileRepository.findById(id);
+        return optionalFile.orElse(null);
     }
 }
